requirements.txt
numpy>=1.24
sounddevice>=0.4.6
PySide6>=6.6
PyYAML>=6.0
pytest>=7.4
